Merge commit 'dcc8db882804b7f8be67dccda40b454834418872' from Apr09 2021

+++ llvm/lib/ObjectYAML/ELFEmitter.cpp
@@ -1359,8 +1359,8 @@ void ELFState<ELFT>::writeSectionContent(
     // Write the address of the function.
     CBA.write<uintX_t>(E.Address, ELFT::TargetEndianness);
     // Write number of BBEntries (number of basic blocks in the function). This
-    // is overriden by the 'NumBlocks' YAML field if specified.
-    uint32_t NumBlocks =
+    // is overridden by the 'NumBlocks' YAML field when specified.
+    uint64_t NumBlocks =
         E.NumBlocks.getValueOr(E.BBEntries ? E.BBEntries->size() : 0);
     SHeader.sh_size += sizeof(uintX_t) + CBA.writeULEB128(NumBlocks);
     // Write all BBEntries.
